+++ BSH/SwaggerClient/SwaggerClient/Classes/Swaggers/Models/Campaign.swift
@@ -17,9 +17,9 @@ public struct Campaign: Codable {
     public var name: String
     public var _description: String
     public var vocabulary: [String]
-    public var userIds: [Double]
+    public var userIds: [Int]
 
-    public init(_id: Int, ownerId: Int, type: CampaignType, name: String, _description: String, vocabulary: [String], userIds: [Double]) {
+    public init(_id: Int, ownerId: Int, type: CampaignType, name: String, _description: String, vocabulary: [String], userIds: [Int]) {
         self._id = _id
         self.ownerId = ownerId
         self.type = type
